+++ Extensions/Revo.Extensions.Notifications/NotificationSerializer.cs
@@ -1,10 +1,29 @@
 ﻿using System;
 using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using Newtonsoft.Json.Serialization;
 
 namespace Revo.Extensions.Notifications
 {
     public class NotificationSerializer : INotificationSerializer
     {
+        private static readonly JsonSerializerSettings JsonSerializerSettings = new JsonSerializerSettings
+        {
+            ContractResolver = new DefaultContractResolver
+            {
+                NamingStrategy = new CamelCaseNamingStrategy
+                {
+                    ProcessDictionaryKeys = false
+                }
+            },
+            Formatting = Formatting.None
+        };
+
+        static NotificationSerializer()
+        {
+            JsonSerializerSettings.Converters.Add(new StringEnumConverter());
+        }
+
         private readonly INotificationTypeCache notificationTypeCache;
 
         public NotificationSerializer(INotificationTypeCache notificationTypeCache)
@@ -17,7 +36,7 @@ namespace Revo.Extensions.Notifications
             SerializedNotification serialized = new SerializedNotification()
             {
                 NotificationClassName = notificationTypeCache.GetNotificationTypeName(notification.GetType()),
-                NotificationJson = JsonConvert.SerializeObject(notification)
+                NotificationJson = JsonConvert.SerializeObject(notification, JsonSerializerSettings)
             };
 
             return serialized;
@@ -26,7 +45,7 @@ namespace Revo.Extensions.Notifications
         public INotification FromJson(SerializedNotification serializedNotification)
         {
             Type notificationType = notificationTypeCache.GetClrNotificationType(serializedNotification.NotificationClassName);
-            return (INotification)JsonConvert.DeserializeObject(serializedNotification.NotificationJson, notificationType);
+            return (INotification)JsonConvert.DeserializeObject(serializedNotification.NotificationJson, notificationType, JsonSerializerSettings);
         }
     }
 }